Add serverspec local tests
By setting the LOCAL_TEST environment variable, one may
switch the serverspec tests to be locally executed on the
current machine. This is handy when you have booted the
module in a vagrant box + puppet provisioning and like to
fire the tests against that setup. By doing that tests can be
executed much faster and therefore are easier to handle.