Change tests, to use a specific testfile_root directory
This changes the tests so that they create files underneath a specific
testfile_root directory so that we must not rely on OS specifc directories,
which are irrelevant for the testing of our config-creation-centric defines.

Further, create_if_empty was renamed to "manage_file", so its clear that
the define will or will-not manage your config_file file ressource.